+++ fiddles/jquery/fiddle-0003-Ajax/app.js
@@ -4,7 +4,7 @@
     $(document).ready(function () {
 
         $(".trigger" ).click(function() {
-            $( ".result" ).load( "data.json" );
+            $(document).load( "data.json" );
         });
 
         $(document).ajaxComplete(function (event, xhr, settings) {