+++ fiddles/phantomjs/fiddle-pageload/test.js
@@ -1,5 +1,0 @@
-var page = require('webpage').create();
-page.open('http://google.com', function() {
-    page.render('google.png');
-    phantom.exit();
-});